% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/code.R
\name{ghdi}
\alias{ghdi}
\title{View Africa GHDI Graph}
\usage{
ghdi(indicator)
}
\arguments{
\item{indicator}{a sting that represent the main inidcators reflecting the Global Human Development Index of Africian countries}
}
\value{
Plot
}
\description{
Function that plot the stock data based on user input. By default the function gives 9 Global Human Development Index indicators of 51 African countries of
}
\details{
The function initially cleans the entered argumets and check their validity. Then update the data file and finally
dispalys the data for the analysis
}
\author{
Seehan
}
